+++ packages/core/src/auth/authorize.ts
@@ -1,6 +1,6 @@
 import 'isomorphic-fetch';
 import isString from 'lodash/isString';
-import { headlessConfig } from '../config';
+import { config } from '../config';
 import { getQueryParam, removeURLParam } from '../utils';
 import { fetchAccessToken } from './client/accessToken';
 
@@ -23,7 +23,7 @@ export async function ensureAuthorization(
 ): Promise<
   true | { redirect?: string | undefined; login?: string | undefined }
 > {
-  const { wpUrl } = headlessConfig();
+  const { wpUrl } = config();
   const { redirectUri, loginPageUri } = options || {};
 
   // Get the authorization code from the URL if it exists